✅ :: Component package more simply
