+++ app/src/main/java/kr/hs/dgsw/smartschool/dui/screen/ButtonScreen.kt
@@ -21,12 +21,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import kr.hs.dgsw.smartschool.components.component.button.ButtonType
-import kr.hs.dgsw.smartschool.components.component.button.DodamLargeRoundedButton
-import kr.hs.dgsw.smartschool.components.component.button.DodamMediumRoundedButton
-import kr.hs.dgsw.smartschool.components.component.button.DodamSmallRoundedButton
-import kr.hs.dgsw.smartschool.components.component.button.IconButton
-import kr.hs.dgsw.smartschool.components.component.button.RadioButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.ButtonType
+import kr.hs.dgsw.smartschool.components.component.basic.button.DodamLargeRoundedButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.DodamMediumRoundedButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.DodamSmallRoundedButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.IconButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.RadioButton
 import kr.hs.dgsw.smartschool.components.theme.DodamColor
 import kr.hs.dgsw.smartschool.components.theme.IcAdd
 import kr.hs.dgsw.smartschool.components.theme.Label3
