+++ dodam-components/src/main/java/kr/hs/dgsw/smartschool/components/component/basic/indicator/PagerIndicator.kt
@@ -1,0 +1,100 @@
+package kr.hs.dgsw.smartschool.components.component.basic.indicator
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.pager.ExperimentalPagerApi
+import com.google.accompanist.pager.PagerState
+import kr.hs.dgsw.smartschool.components.theme.DodamTheme
+
+@OptIn(ExperimentalPagerApi::class)
+@Composable
+fun PagerIndicator(
+    modifier: Modifier = Modifier,
+    pagerState: PagerState,
+    indicatorCount: Int,
+    indicatorSize: Dp = 6.dp,
+    indicatorShape: Shape = CircleShape,
+    space: Dp = 3.dp,
+    activeColor: Color = DodamTheme.color.MainColor,
+    inActiveColor: Color = DodamTheme.color.White,
+    onClick: ((Int) -> Unit)? = null
+) {
+
+    val listState = rememberLazyListState()
+
+    val totalWidth: Dp = indicatorSize * indicatorCount + space * (indicatorCount - 1)
+    val widthInPx = LocalDensity.current.run { indicatorSize.toPx() }
+
+    val currentItem by remember {
+        derivedStateOf {
+            pagerState.currentPage
+        }
+    }
+
+    val itemCount = pagerState.pageCount
+
+    LaunchedEffect(key1 = currentItem) {
+        val viewportSize = listState.layoutInfo.viewportSize
+        listState.animateScrollToItem(
+            currentItem,
+            (widthInPx / 2 - viewportSize.width / 2).toInt()
+        )
+    }
+
+    LazyRow(
+        modifier = modifier.width(totalWidth),
+        state = listState,
+        contentPadding = PaddingValues(vertical = space),
+        horizontalArrangement = Arrangement.spacedBy(space),
+        userScrollEnabled = false
+    ) {
+
+        items(itemCount) { index ->
+
+            val isSelected = (index == currentItem)
+
+            Box(
+                modifier = Modifier
+                    .graphicsLayer {
+                        scaleX = 1f
+                        scaleY = 1f
+                    }
+                    .clip(indicatorShape)
+                    .size(indicatorSize)
+                    .background(
+                        if (isSelected) activeColor else inActiveColor,
+                        indicatorShape
+                    )
+                    .then(
+                        if (onClick != null) {
+                            Modifier
+                                .clickable {
+                                    onClick.invoke(index)
+                                }
+                        } else Modifier
+                    )
+            )
+        }
+    }
+}
